requirements.txt
crewai==0.36.1
crewai_tools==0.4.8
fastapi==0.111.0
langchain_community==0.2.7
langchain_google_genai==1.0.7
llama_index==0.10.55
motor==3.5.1
pydantic==2.8.2
python-dotenv==1.0.1
